+++ pytest_helm_charts/utils.py
@@ -4,7 +4,7 @@ import time
 from typing import Dict, Any, List, TypeVar, Callable, Type, Optional
 
 import pykube.exceptions
-from pykube import HTTPClient, Deployment
+from pykube import HTTPClient
 
 YamlDict = Dict[str, Any]
 
@@ -77,59 +77,6 @@ def wait_for_namespaced_objects_condition(
         raise TimeoutError(f"Error waiting for object of type {obj_type} to match the condition.")
 
     return matching_objs
-
-
-def _deployment_running(deploy: Deployment) -> bool:
-    complete = (
-        "status" in deploy.obj
-        and "availableReplicas" in deploy.obj["status"]
-        and "observedGeneration" in deploy.obj["status"]
-        and "updatedReplicas" in deploy.obj["status"]
-        and int(deploy.obj["status"]["observedGeneration"]) >= int(deploy.obj["metadata"]["generation"])
-        and deploy.replicas == int(deploy.obj["status"]["updatedReplicas"])
-        and deploy.replicas == int(deploy.obj["status"]["availableReplicas"])
-    )
-    return complete
-
-
-def wait_for_deployments_to_run(
-    kube_client: HTTPClient,
-    deployment_names: List[str],
-    deployments_namespace: str,
-    timeout_sec: int,
-    missing_ok: bool = True,
-) -> List[Deployment]:
-    """
-    Block until all the Deployments are running or timeout is reached.
-
-    Args:
-        kube_client: client to use to connect to the k8s cluster
-        deployment_names: a list of Deployment names to check
-        deployments_namespace: namespace where all the Deployments are created (single namespace for all resources)
-        timeout_sec: timeout for the call
-        missing_ok: when `True`, the function ignores that some of the objects listed in the `deployment_names`
-            don't exist in k8s API and waits for them to show up; when `False`, an
-            [ObjectNotFound](pykube.exceptions.ObjectDoesNotExist) exception is raised.
-
-    Returns:
-        The list of Deployment resources with all the objects listed in `deployment_names` included.
-
-    Raises:
-        TimeoutError: when timeout is reached.
-        pykube.exceptions.ObjectDoesNotExist: when `missing_ok == False` and one of the objects
-            listed in `deployment_names` can't be found in k8s API
-
-    """
-    result = wait_for_namespaced_objects_condition(
-        kube_client,
-        Deployment,
-        deployment_names,
-        deployments_namespace,
-        _deployment_running,
-        timeout_sec,
-        missing_ok,
-    )
-    return result
 
 
 def _stateful_set_ready(sts: pykube.StatefulSet) -> bool:
